Update 5
Scrapped the PixiUI files since I didn't think I actually needed them.

Switched the background for the credits and the title screen to make the instructions easier to read.

Added button functionality, though the start button functionality does need to be fixed.

Removed one of the fonts. I plan on getting a new one to replace it.

Added credits scene. Crediting is still a work in progress.

Moved scene transition-related code to a new JS file.